backend/requirements.txt
Flask>=3.0.0,<4.0.0
numpy>=1.20.0,<2.0.0
pymongo[srv]>=4.0.0,<5.0.0
python-dotenv>=0.19.0,<1.0.0